refactor[#40] : 기존 entities/article 에 존재하던 메소드들 파일로 이동
